Create Pythontutor_008_"Laces"
The Task Of The "Laces"
Condition
Shoe factory is going to start producing a luxury model of Shoe. The holes for lacing will be located in two rows, the distance between the rows is a, and the distance between holes in row b. The number of holes in each row is N. Lace-up must occur the elite way “upward, horizontally to a different number, upward, horizontally, etc.” (see picture). In addition to laces could tie elite bow, the length of the free end of the cord l. What should be the length of the lace for these boots?

The program receives input four positive integers a, b, l and N in that order, and should print a single number - the desired length of the lace.



In all tasks read input using input() and print the answer using print().